requirements.txt
PySide6
python-docx
python-dateutil
